chore: components renamed
